requirements.txt
kiteconnect==4.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.1.0
numpy>=1.25.0
pyyaml==6.0.1
rich==13.7.0
python-dateutil==2.8.2
